- Fixing build errors again...

+++ event_horizon/graphics/audio/audio_file.cpp
@@ -304,7 +304,7 @@ bool AudioFile<T>::decodeWaveFile (std::vector<uint8_t>& fileData)
     }
     
     // check header data is consistent
-    if ((numBytesPerSecond != (numChannels * sampleRate * bitDepth) / 8) || (numBytesPerBlock != (numChannels * numBytesPerSample)))
+    if ((numBytesPerSecond != (numChannels * static_cast<int16_t>(sampleRate * bitDepth)) / 8) || (numBytesPerBlock != (numChannels * numBytesPerSample)))
     {
         std::cout << "ERROR: the header data in this WAV file seems to be inconsistent" << std::endl;
         return false;
